+++ subpackages/cli/source/odood/cli/app.d
@@ -27,6 +27,7 @@ private import odood.cli.commands.log: CommandLogView;
 private import odood.cli.commands.script: CommandScript;
 private import odood.cli.commands.psql: CommandPSQL;
 private import odood.cli.commands.ci: CommandCi;
+private import odood.cli.commands.info: CommandInfo;
 
 
 /** Class that represents main OdoodProgram
@@ -52,6 +53,7 @@ class App: OdoodProgram {
         this.add(new CommandScript());
         this.add(new CommandPSQL());
         this.add(new CommandCi());
+        this.add(new CommandInfo());
 
         // shortcuts
         this.add(new CommandServerStart());